--- a/CHANGELOG.rst
+++ b/CHANGELOG.rst
@@ -8,6 +8,7 @@ Added
 Changed
 ~~~~~~~~
 * Introduce github actions for tests instead of TravisCI
+* Replace pygeodesy with pyproj for speed-up
 Deprecated
 ~~~~~~~~~~~~
 Removed
@@ -19,11 +20,6 @@ Fixed
 Security
 ~~~~~~~~~
 
-v0.1.8
-------------------------
-Changed
-~~~~~~~~
-* Replace pygeodesy with pyproj for speed-up
 v0.1.7
 ------------------------
 Fixed