--- a/requirements.txt
+++ b/requirements.txt
@@ -1,4 +1,3 @@
-pygeodesy
 pyproj
 aerofiles
 beautifulsoup4