--- a/setup.py
+++ b/setup.py
@@ -16,5 +16,6 @@ setup(
     install_requires=[
         'aerofiles>=0.4.1',
         'beautifulsoup4>=4.6.0'
+        'pyproj>=3.4.1'
     ]
 )