I know that `opensoar` is not actively developed at the moment so don't feel pressured to look into this PR. 

I was always wondering why `opensoar` is relatively slow. As Christmas is approaching and I had a little spare time, I did some profiling of the code and found the major bottleneck to be related to `pygeodesy`. By replacing it with `pyproj`, I could achieve a major speed-up of up to a factor of 15, without effects on the results for `distance` and `bearing`. No changes were observable for the `pysoar` analysis, too.

I did some timing with `pygeodesy` and  `pyproj` for racing and AAT task (average over 10 calls of `Trip`):

|                         | Racing  | AAT        | 
| -----------          | ----------- | ----------- |
|  `pyproj`         | 0.33 s    | 2.14 s    | 
| `pygeodesy`  | 4.9 s      | 30.7 s    |