@GliderGeek 2022-12-28
Thanks @jkretz for this contribution! The numbers look very promising and i don't have a specific reason to stick to `pygeodesy`, so this is very welcome.

Hopefully I'll soon find time to review this. I recently started on a new version of pysoar (mainly maintenance and ci), so fixing this first and shipping increased performance makes for a nice combination :)