@GliderGeek 2023-01-05
@jkretz : i just added github action to make sure tests are run on each pull request. could you update your pull request such that it takes in these changes?