@jkretz 2023-01-05
I rebased my branch. 

Should we maybe also add some kind of linting to the github actions? I would have a workflow for python linting lying around somewhere using the super-linter.