@jkretz 2023-01-05
Of course using a linter might end up in some refactoring of the code