@GliderGeek 2023-01-05
wrt linting: was it difficult for you to add your code to this codebase and would a linter help? 

if not, i prefer not to add tooling that adds complexity and potential false positives