@jkretz 2023-01-06
I agree with that linting might not be necessary for such a small project.

I will get rid of the overlapping code segments.