@GliderGeek 2023-01-09
Thanks for the great contribution @jkretz! it will be used in the upcoming pysoar release: https://github.com/GliderGeek/PySoar/pull/163